backend/requirements.txt
uvicorn
python-dotenv
fastapi
openai